. . . testing . . .

+++ users-service/src/main/java/com/fitnesstracker/users/presentationlayer/UserController.java
@@ -3,6 +3,7 @@ package com.fitnesstracker.users.presentationlayer;
 
 import com.fitnesstracker.users.businesslayer.UserService;
 import com.fitnesstracker.users.utils.exceptions.InvalidInputException;
+import com.fitnesstracker.users.utils.exceptions.NotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -29,7 +30,11 @@ public class UserController {
         if (userId.length() != UUID_LENGTH) {
             throw new InvalidInputException("Invalid userId provided: " + userId);
         }
-        return ResponseEntity.ok().body(userService.getUserByUserId(userId));
+        UserResponseModel user = userService.getUserByUserId(userId);
+        if (user == null && userId.length() == UUID_LENGTH) {
+            throw new NotFoundException("User not found: " + userId);
+        }
+        return ResponseEntity.ok().body(user);
     }
 
     @PostMapping()
@@ -39,6 +44,10 @@ public class UserController {
 
     @PutMapping("/{userId}")
     public ResponseEntity<UserResponseModel> updateUser(@RequestBody UserRequestModel userRequestModel, @PathVariable String userId) {
+        UserResponseModel updatedUser = userService.getUserByUserId(userId);
+        if (updatedUser == null && userId.length() == UUID_LENGTH) {
+            throw new NotFoundException("User not found: " + userId);
+        }
         if (userId.length() != UUID_LENGTH) {
             throw new InvalidInputException("Invalid userId provided: " + userId);
         }
@@ -49,6 +58,10 @@ public class UserController {
     public ResponseEntity<Void> deleteUser(@PathVariable String userId) {
         if (userId.length() != UUID_LENGTH) {
             throw new InvalidInputException("Invalid userId provided: " + userId);
+        }
+        UserResponseModel user = userService.getUserByUserId(userId);
+        if (user == null && userId.length() == UUID_LENGTH) {
+            throw new NotFoundException("User not found: " + userId);
         }
         userService.deleteUser(userId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
